PART6-DJANGO_P2/requirements.txt
absl-py==0.13.0
argon2-cffi==20.1.0
asgiref==3.5.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==4.0.2
djangorestframework==3.13.1
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==2.8
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
json5==0.9.5
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.3
notebook==6.1.5
oauthlib==3.1.1
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.7
pandas==1.2.0
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.3.1
ply==3.11
portpicker==1.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.20
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.1.1
QtPy==1.10.0
regex==2020.11.13
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.0
scipy==1.6.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
sqlparse==0.4.2
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzdata==2021.5
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
